@@ -41,7 +41,9 @@ class ShareExperienceForm(forms.Form):
     other.group = 2
 
     # sharing options
-    viewable = forms.BooleanField(label = "Share on AutSPACE website", required=False)
+    viewable = forms.BooleanField(label = "Share on AutSPACE website", 
+                                  required=False,
+                                  widget=forms.CheckboxInput(attrs={'id':'shareOnAutSPACEs'}))
     viewable.group = 3
     research = forms.BooleanField(label = "Share for research", required=False)
     research.group = 3
@@ -115,6 +117,9 @@ class ModerateExperienceForm(forms.Form):
                                                               'rows':'4',
                                                               'class':'form-control'}))
     moderation_comments.group = "hidden"
+
+    moderation_prior = forms.CharField(widget=forms.HiddenInput())
+    moderation_prior.group = "hidden"
 
     def __init__(self, *args, **kwargs):
         """ Disable free text fields to the moderator, or disable all fields if in 'read only' mode"""
